% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_function.R
\name{apply_cl_cm}
\alias{apply_cl_cm}
\title{apply_cl_cm}
\usage{
apply_cl_cm(
  data,
  vars_bolean = TRUE,
  vars_ordered = c(1:50),
  data_selected_order = TRUE,
  save_ordered_data = "path/to/save/data.csv",
  column_mapping = TRUE,
  config_folder_path = "/config_folder_path/column_mapping.rds"
)
}
\arguments{
\item{data}{a dataframe or a tibble}

\item{vars_ordered}{selecting the variables in the order you want. it is similiar to dplyr::select('var1', 'var2').}

\item{data_selected_order}{a bolean TRUE or FALSE. If TRUE then dataset obtained in vars_ordered will be saved into a folder of your choice.}

\item{save_ordered_data}{the path folder where the data is saved}

\item{column_mapping}{a bolean TRUE OR FALSE. If TRUE, then a column mapping tibble with 2 variables - 'provided' and 'standard'.
provided variable stores all the variables of data passed into this function, stored as rows; standard variable contains the same variables as in provided column but
all variables are transformed from CAPITAL letters variables to lower letters. e.g. PSEUDO_ID -> pseudo_id}

\item{config_folder_path}{the path of your choice where the column_mapping tibble (and other tibbles that form the process of recoding levels of variables) are stored}
}
\value{

}
\description{
apply_cl_cm
}
