% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_function.R
\name{cfg_write}
\alias{cfg_write}
\title{cfg_write}
\usage{
cfg_write(
  provided = c(1:2),
  standard = c(1:2),
  table = name_of_a_table,
  path = "path/to/save"
)
}
\arguments{
\item{provided}{categorical levels of the variable that are provided in the original dataset}

\item{standard}{levels recoded to new categories}

\item{table}{the name of the table to which new table should be saved}

\item{path}{the path of the folder to which the table should be saved}
}
\value{

}
\description{
cfg_write
}
