\name{county.fips}
\docType{data}
\alias{county.fips}
\title{
FIPS county codes for US County Map
}
\description{
A database matching FIPS codes to maps package county and state names.
}
\usage{data(county.fips)}
\format{
A list with 2 components, namely "fips" and "polyname", containing the
FIPS number and respective state or county polygon name.
}
\seealso{
\code{\link{state.fips}}
}
\keyword{datasets}
