% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boundaries.R
\name{get_coastline}
\alias{get_coastline}
\title{Get Ireland Coastline}
\usage{
get_coastline(year = 2016, resolution = "03")
}
\arguments{
\item{year}{Numeric. Reference year. Default 2016.}

\item{resolution}{Character. Resolution code. Default "03".}
}
\value{
An sf object of coastlines.
}
\description{
Fetch coastline geometry for intersection with county boundaries.
}
