\name{internal}
\alias{internal}
\alias{makepoly}
\alias{mapgetg}
\alias{mapgetl}
\alias{mapname}
\alias{mapthin}
\alias{maptype}
\title{Internally Required Functions}

\usage{
makepoly(xy, gonsize, keep)
mapgetg(database, gons, fill, xlim, ylim)
mapgetl(database, lines, xlim, ylim, fill)
mapname(database, patterns, exact)
mapthin(xy, delta, symmetric)
maptype(database)
}

\description{
These functions are called internally and will generally not be
required by the user.
}

\keyword{internal}
