\name{ozone}
\docType{data}
\alias{ozone}
\alias{votes.repub}
\alias{unemp}
\title{Sample datasets}
\usage{
data(ozone)
data(unemp)
data(votes.repub)
}
\description{Datasets used to illustrate map functions.

\code{ozone} contains the median of daily maxima ozone concentration in 41 US cities for June 1974 through August 1974. Concentrations are in parts per billion (ppb).

\code{unemp} Has population and unemployment percentage for US counties.

\code{votes.repub} contains the percentage republican votes in the 1900 election.
}
\references{
Cleveland, W.S., Kleiner, B., McRae, J.E., Warner, J.L., and Pasceri, P.E. ,
"The Analysis of Ground-Level Ozone Data from New Jersey, New York, Connecticut, and Massachusetts: Data Quality Assessment and Temporal and Geographical Properties", \emph{Bell Laboratories Memorandum, 1975}.


}
\keyword{datasets}
