\name{polygon}
\alias{apply.polygon}
\alias{area.polygon}
\alias{centroid.polygon}
\alias{as.matrix.polygon}
\alias{closed.polygon}
\alias{in.polygon}
\alias{in.one.polygon}
\alias{num.polygons}
\alias{sub.polygon}
\title{Polygon functions}
\usage{
area.polygon(p)
centroid.polygon(p)
\method{as.matrix}{polygon}(x, ...)
closed.polygon(p)
in.one.polygon(p, x)
in.polygon(p, x)
num.polygons(p)
sub.polygon(p, i)
}
\description{
These functions are called internally and will generally not be
required by the user.
}
\keyword{internal}
