\name{us.cities}
\docType{data}
\alias{us.cities}
\title{
Database of US cities
}

\description{
This database is of us cities of population greater than about 40,000.
Also included are state capitals of any population size.
}
\format{
A list with 6 components, namely "name", "country.etc", "pop",  "lat",
"long", and "capital", containing the city name, the state abbreviation,
approximate population (as at January 2006), latitude, longitude and
capital status indication (0 for non-capital, 1 for capital, 2 for state
capital.
}
\section{NOTE}{
Some of the city names may be out of date.  Please send
any corrections to the package maintainer.
}
\seealso{
\code{\link{map.cities}}
}
\keyword{datasets}
