% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sea_distance.R
\name{compute_sea_distance}
\alias{compute_sea_distance}
\title{Compute Sea Distance for Counties}
\usage{
compute_sea_distance(counties, coastline)
}
\arguments{
\item{counties}{An sf object of county polygons.}

\item{coastline}{An sf linestring of coastlines.}
}
\value{
The input sf object with an added \code{sea_distance} integer column.
}
\description{
Classify counties by their graph distance from the sea.
Distance 0 = touches the sea, distance 1 = neighbour of a coastal county, etc.
}
