\name{world.cities}
\docType{data}
\alias{world.cities}
\title{
Database of world cities
}

\description{
This database is primarily of world cities of population greater than
about 40,000.  Also included are capital cities of any population size, and
many smaller towns.
}
\usage{data(world.cities)}
\format{
A list with 6 components, namely "name", "country.etc", "pop", "lat",
"long", and "capital", containing the city name, the country name,
approximate population (as at January 2006), latitude, longitude and
capital status indication (0 for non-capital, 1 for capital, 2 for China
Municipalities, and 3 for China Provincial capitals)
}
\section{NOTE}{
Some of the country names and city names may be out of date.  Please send
any corrections to the package maintainer.
}
\seealso{
\code{\link{map.cities}}
}
\keyword{datasets}
