\name{world.cities}
\docType{data}
\alias{world.cities}
\title{
Database of world cities
}

\description{
This database is primarily of world cities of population greater than
100,000.  Also included are capital cities of any population size, and
some local additions.
}
\format{
A list with 6 components, namely "name", "country.etc", "lat", "long",
"pop",  and "capital", containing the city name (in uppercase for
capital cities), the country name, latitude, longitude, approximate
population, and capital status indication (0 for non-capital, 1 for
capital, 2 for China Municipalities, and 3 for China Provincial
capitals)
}
\section{NOTE}{
Some of the country names and city names are out of date.  Please send
any corrections to the package maintainer.
}
\seealso{
\code{\link{map.cities}}
}
\keyword{datasets}
